+++ src/sources/GeoJSON.js
@@ -1,14 +1,18 @@
 import * as rsys from '../client/rsys';
 import Dataframe from '../renderer/Dataframe';
-import Metadata from './GeoJSONMetadata';
-import CartoValidationError, { CartoValidationErrorTypes } from '../../src/errors/carto-validation-error';
-import CartoRuntimeError, { CartoRuntimeErrorTypes } from '../../src/errors/carto-runtime-error';
+import { DEFAULT_ID_PROPERTY } from '../renderer/Metadata';
+
+import CartoValidationError, { CartoValidationErrorTypes } from '../errors/carto-validation-error';
+
 import util from '../utils/util';
 import Base from './Base';
-import schema from '../renderer/schema';
-import { GEOMETRY_TYPE } from '../utils/geometry';
 
-const SAMPLE_TARGET_SIZE = 1000;
+import GeoJSONGeometryTransformer from './geojson/GeoJSONGeometryTransformer';
+import { GeoJSONGeometryType, dataframeGeometryType } from './geojson/GeoJSONGeometryType';
+import { GeoJSONMetadataBuilder } from './geojson/GeoJSOMetadataBuilder';
+
+const DATAFRAME_PADDING = 1024;
+const SAMPLE_SIZE_FOR_CENTER = 10;
 
 export default class GeoJSON extends Base {
     /**
@@ -40,30 +44,101 @@ export default class GeoJSON extends Base {
      */
     constructor (data, options = {}) {
         super();
+
+        this._initializeData(data);
+        this._initializeMetadataHelpers(options);
+    }
+
+    _initializeData (data) {
         this._checkData(data);
 
-        this._type = ''; // Point, LineString, MultiLineString, Polygon, MultiPolygon
-        this._categoryStringToIDMap = {};
-        this._numCategories = 0;
-        this._numFields = new Set();
-        this._catFields = new Set();
-        this._dateFields = new Set();
+        this._data = data;
+        this._features = this._featuresFromData();
+        this._geomType = this._getGeomType();
+        this._webMercatorCenter = this._getWebMercatorCoordsCenter();
+        this._geometryTransformer = new GeoJSONGeometryTransformer(this._webMercatorCenter);
+    }
+
+    _initializeMetadataHelpers (options) {
         this._providedDateColumns = new Set(options.dateColumns);
-        this._properties = {};
         this._boundColumns = new Set();
 
-        this._data = data;
-        if (data.type === 'FeatureCollection') {
-            this._features = data.features;
-        } else if (data.type === 'Feature') {
-            this._features = [data];
+        this._metadataBuilder = new GeoJSONMetadataBuilder(this._providedDateColumns, this._boundColumns);
+    }
+
+    /**
+     * Check geojson data is a proper object
+     */
+    _checkData (data) {
+        if (util.isUndefined(data)) {
+            throw new CartoValidationError('\'data\'', CartoValidationErrorTypes.MISSING_REQUIRED);
+        }
+        if (!util.isObject(data)) {
+            throw new CartoValidationError('\'data\' property must be an object.',
+                CartoValidationErrorTypes.INCORRECT_TYPE);
+        }
+    }
+
+    /**
+     * Get initialized Features from data
+     */
+    _featuresFromData () {
+        let features;
+
+        const dataType = this._data.type;
+        if (dataType === 'FeatureCollection') {
+            features = this._data.features;
+        } else if (dataType === 'Feature') {
+            features = [this._data];
         } else {
             throw new CartoValidationError('\'data\' property must be a GeoJSON object.', CartoValidationErrorTypes.INCORRECT_VALUE);
         }
 
-        this._features = this._initializeFeatureProperties(this._features);
+        this._initializePropertiesIn(features);
+        return features;
+    }
 
-        this._setCoordinatesCenter();
+    /**
+     * Get the GeoJSONGeometryType from the first Feature
+     */
+    _getGeomType () {
+        return this._features[0].geometry.type;
+    }
+
+    /**
+     * Get an estimated center in WebMercator coordinates, based on a sample of the Features
+     * with SAMPLE_SIZE_FOR_CENTER
+     */
+    _getWebMercatorCoordsCenter () {
+        let x = 0;
+        let y = 0;
+        let nPoints = 0;
+        this._fetchFeatureGeometry({ sample: SAMPLE_SIZE_FOR_CENTER }, (_, geometry) => {
+            const samplePoint = this._samplePoint(geometry);
+            const sampleXY = util.projectToWebMercator({ lng: samplePoint[0], lat: samplePoint[1] });
+            x += sampleXY.x;
+            y += sampleXY.y;
+            nPoints += 1;
+        });
+
+        if (nPoints > 1) {
+            x /= nPoints;
+            y /= nPoints;
+        }
+
+        return { x, y };
+    }
+
+    /**
+     * Get the dataframe center in a local RSys.
+     */
+    _getDataframeCenter () {
+        const targetReferenceSystem = {
+            scale: util.WM_R,
+            center: { x: 0, y: 0 }
+        };
+        const { x: worldX, y: worldY } = this._webMercatorCenter;
+        return rsys.wToR(worldX, worldY, targetReferenceSystem);
     }
 
     bindLayer (addDataframe) {
@@ -77,26 +152,36 @@ export default class GeoJSON extends Base {
     requestData () {
         if (this._dataframe) {
             const newProperties = this._decodeUnboundProperties();
-            this._dataframe.addProperties(newProperties);
-            Object.keys(newProperties).forEach(propertyName => {
+            this._dataframe.addProperties();
+
+            const newPropertyNames = Object.keys(newProperties);
+            newPropertyNames.forEach(propertyName => {
                 this._boundColumns.add(propertyName);
             });
-            return Promise.resolve(Object.keys(newProperties).length > 0);
+
+            return Promise.resolve(newPropertyNames.length > 0);
         }
+
+        this._dataframe = this._createNewDataframe();
+        return Promise.resolve(true);
+    }
+
+    _createNewDataframe () {
         const dataframe = new Dataframe({
             active: true,
-            center: this._dataframeCenter,
+            center: this._getDataframeCenter(),
             geom: this._decodeGeometry(),
             properties: this._decodeUnboundProperties(),
             scale: 1,
             size: this._features.length,
-            type: this._getDataframeType(this._type),
+            type: dataframeGeometryType(this._geomType),
             metadata: this._metadata
         });
+
         this._boundColumns = new Set(Object.keys(dataframe.properties));
-        this._dataframe = dataframe;
         this._addDataframe(dataframe);
-        return Promise.resolve(true);
+
+        return dataframe;
     }
 
     requiresNewMetadata () {
@@ -107,214 +192,61 @@ export default class GeoJSON extends Base {
         return new GeoJSON(this._data, { dateColumns: Array.from(this._providedDateColumns) });
     }
 
-    _checkData (data) {
-        if (util.isUndefined(data)) {
-            throw new CartoValidationError('\'data\'', CartoValidationErrorTypes.MISSING_REQUIRED);
-        }
-        if (!util.isObject(data)) {
-            throw new CartoValidationError('\'data\' property must be an object.', CartoValidationErrorTypes.INCORRECT_TYPE);
-        }
-    }
-
-    _initializeFeatureProperties (features) {
+    _initializePropertiesIn (features) {
         for (let i = 0; i < features.length; i++) {
             features[i].properties = features[i].properties || {};
         }
-
-        return features;
     }
 
     _computeMetadata (viz) {
-        const sample = [];
-        this._addNumericColumnField('cartodb_id');
-
-        const featureCount = this._features.length;
-        const requiredColumns = new Set(Object.keys(schema.simplify(viz.getMinimumNeededSchema())));
-        for (let i = 0; i < this._features.length; i++) {
-            const properties = this._features[i].properties;
-            const keys = Object.keys(properties);
-            for (let j = 0, len = keys.length; j < len; j++) {
-                const name = keys[j];
-                if (!requiredColumns.has(name) || this._boundColumns.has(name)) {
-                    continue;
-                }
-                const value = properties[name];
-                this._addPropertyToMetadata(name, value);
-            }
-            this._sampleFeatureOnMetadata(properties, sample, this._features.length);
-        }
-
-        this._numFields.forEach(name => {
-            const property = this._properties[name];
-            property.avg = property.sum / property.count;
-        });
-
-        let geomType = '';
-        if (featureCount > 0) {
-            // Set the geomType of the first feature to the metadata
-            geomType = this._getDataframeType(this._features[0].geometry.type);
-        }
-        const idProperty = 'cartodb_id';
-
-        this._metadata = new Metadata({ properties: this._properties, featureCount, sample, geomType, idProperty });
-        this._metadata.setCodecs();
+        this._metadata = this._metadataBuilder.buildFrom(viz, this._features);
         return this._metadata;
     }
 
-    _sampleFeatureOnMetadata (properties, sample, featureCount) {
-        if (featureCount > SAMPLE_TARGET_SIZE) {
-            const sampling = SAMPLE_TARGET_SIZE / featureCount;
-            if (Math.random() > sampling) {
-                return;
-            }
-        }
-        sample.push(properties);
-    }
-
-    _addNumericPropertyToMetadata (propertyName, value) {
-        if (this._catFields.has(propertyName) || this._dateFields.has(propertyName)) {
-            throw new CartoValidationError(
-                `Unsupported GeoJSON: the property '${propertyName}' has different types in different features.`,
-                CartoValidationErrorTypes.INCORRECT_TYPE
-            );
-        }
-        this._addNumericColumnField(propertyName);
-        const property = this._properties[propertyName];
-        property.min = Math.min(property.min, value);
-        property.max = Math.max(property.max, value);
-        property.sum += value;
-    }
-
-    _addNumericColumnField (propertyName) {
-        if (!this._numFields.has(propertyName)) {
-            this._numFields.add(propertyName);
-            this._properties[propertyName] = {
-                type: 'number',
-                min: Number.POSITIVE_INFINITY,
-                max: Number.NEGATIVE_INFINITY,
-                avg: Number.NaN,
-                sum: 0,
-                count: 0
-            };
-        }
-    }
-
-    _addDatePropertyToMetadata (propertyName, value) {
-        if (this._catFields.has(propertyName) || this._numFields.has(propertyName)) {
-            throw new CartoRuntimeError(
-                `Unsupported GeoJSON: the property '${propertyName}' has different types in different features.`,
-                CartoRuntimeErrorTypes.NOT_SUPPORTED
-            );
-        }
-        this._addDateColumnField(propertyName);
-        const column = this._properties[propertyName];
-        const dateValue = util.castDate(value);
-        column.min = column.min ? util.castDate(Math.min(column.min, dateValue)) : dateValue;
-        column.max = column.max ? util.castDate(Math.max(column.max, dateValue)) : dateValue;
-        column.sum += value;
-        column.count++;
-    }
-
-    _addDateColumnField (propertyName) {
-        if (!this._dateFields.has(propertyName)) {
-            this._dateFields.add(propertyName);
-            this._properties[propertyName] = {
-                type: 'date',
-                min: null,
-                max: null,
-                avg: null,
-                sum: 0,
-                count: 0
-            };
-        }
-    }
-
-    _addPropertyToMetadata (propertyName, value) {
-        if (this._providedDateColumns.has(propertyName)) {
-            return this._addDatePropertyToMetadata(propertyName, value);
-        }
-        if (Number.isFinite(value)) {
-            return this._addNumericPropertyToMetadata(propertyName, value);
-        }
-        if (value === null) {
-            return;
-        }
-        this._addCategoryPropertyToMetadata(propertyName, value);
-    }
-
-    _addCategoryPropertyToMetadata (propertyName, value) {
-        if (this._numFields.has(propertyName) || this._dateFields.has(propertyName)) {
-            throw new CartoRuntimeError(
-                `Unsupported GeoJSON: the property '${propertyName}' has different types in different features.`,
-                CartoRuntimeErrorTypes.NOT_SUPPORTED
-            );
-        }
-        if (!this._catFields.has(propertyName)) {
-            this._catFields.add(propertyName);
-            this._properties[propertyName] = {
-                type: 'category',
-                categories: []
-            };
-        }
-        const property = this._properties[propertyName];
-        const cat = property.categories.find(cat => cat.name === value);
-        if (cat) {
-            cat.frequency++;
-        } else {
-            property.categories.push({ name: value, frequency: 1 });
-        }
-    }
-
     _decodeUnboundProperties () {
-        const properties = {};
-        [...this._numFields].concat([...this._catFields]).concat([...this._dateFields]).map(name => {
-            if (this._boundColumns.has(name)) {
-                return;
-            }
-            // The dataframe expects to have a padding of 1024, adding 1024 empty values assures this condition is met
-            properties[name] = new Float32Array(this._features.length + 1024);
-        });
-
-        const catFields = [...this._catFields].filter(name => !this._boundColumns.has(name));
-        const numFields = [...this._numFields].filter(name => !this._boundColumns.has(name));
-        const dateFields = [...this._dateFields].filter(name => !this._boundColumns.has(name));
-        const fields = [...catFields, ...numFields, ...dateFields];
+        const unboundProperties = this._getUnboundProperties();
+        const unboundFieldNames = Object.keys(unboundProperties);
 
         for (let i = 0; i < this._features.length; i++) {
             const f = this._features[i];
-            fields.forEach(name => {
-                if (name === 'cartodb_id' && !Number.isFinite(f.properties.cartodb_id)) {
-                    // Using negative ids for GeoJSON features
-                    f.properties.cartodb_id = -i;
+            unboundFieldNames.forEach(name => {
+                if (name === DEFAULT_ID_PROPERTY && !Number.isFinite(f.properties[DEFAULT_ID_PROPERTY])) {
+                    f.properties[DEFAULT_ID_PROPERTY] = -i; // using negative ids for GeoJSON features
                 }
                 // note that GeoJSON does not support multi-value properties
-                properties[name][i] = this._metadata.codec(name).sourceToInternal(this._metadata, f.properties[name]);
+                unboundProperties[name][i] = this._metadata.codec(name).sourceToInternal(this._metadata, f.properties[name]);
             });
         }
+
+        return unboundProperties;
+    }
+
+    _getUnboundProperties () {
+        const allFields = this._metadataBuilder.getCurrentFields();
+
+        const properties = {};
+        allFields.forEach(name => {
+            if (this._boundColumns.has(name)) {
+                return;
+            }
+            // The dataframe expects to have a padding, adding DATAFRAME_PADDING empty values assures this condition is met
+            properties[name] = new Float32Array(this._features.length + DATAFRAME_PADDING);
+        });
+
         return properties;
     }
 
-    _getDataframeType (type) {
-        switch (type) {
-            case 'Point':
-                return GEOMETRY_TYPE.POINT;
-            case 'LineString':
-            case 'MultiLineString':
-                return GEOMETRY_TYPE.LINE;
-            case 'Polygon':
-            case 'MultiPolygon':
-                return GEOMETRY_TYPE.POLYGON;
-            default:
-                return '';
-        }
-    }
-
+    /**
+     * Take the geometry from every feature and invoke the callback with it.
+     * If custom options are specified, a 'sample' can be used to skip every 'X' features
+     */
     _fetchFeatureGeometry (options = {}, callback) {
         let geometry = null;
+        const DEFAULT_SAMPLE = 1; // every feature
         const numFeatures = this._features.length;
-        const incr = options.sample ? Math.max(1, Math.floor(numFeatures / options.sample)) : 1;
+        const increment = options.sample ? Math.max(1, Math.floor(numFeatures / options.sample)) : DEFAULT_SAMPLE;
 
-        for (let i = 0; i < numFeatures; i += incr) {
+        for (let i = 0; i < numFeatures; i += increment) {
             const feature = this._features[i];
             if (feature.type === 'Feature') {
                 callback(i, feature.geometry);
@@ -324,7 +256,7 @@ export default class GeoJSON extends Base {
     }
 
     _allocGeometry () {
-        if (this._type === 'Point') {
+        if (this._geomType === 'Point') {
             return new Float32Array(this._features.length * 6);
         }
         return ([]);
@@ -336,31 +268,31 @@ export default class GeoJSON extends Base {
         this._fetchFeatureGeometry({}, (i, geometry) => {
             const type = geometry.type;
             const coordinates = geometry.coordinates;
-            if (this._type !== type) {
+            if (this._geomType !== type) {
                 throw new CartoValidationError(
-                    `multiple geometry types not supported: found '${type}' instead of '${this._type}'.`,
+                    `multiple geometry types not supported: found '${type}' instead of '${this._geomType}'.`,
                     CartoValidationErrorTypes.INCORRECT_TYPE
                 );
             }
-            if (type === 'Point') {
-                const point = this._computePointGeometry(coordinates);
+            if (type === GeoJSONGeometryType.POINT) {
+                const point = this._geometryTransformer.computePoint(coordinates);
                 geometries[6 * i + 0] = point.x;
                 geometries[6 * i + 1] = point.y;
                 geometries[6 * i + 2] = point.x;
                 geometries[6 * i + 3] = point.y;
                 geometries[6 * i + 4] = point.x;
                 geometries[6 * i + 5] = point.y;
-            } else if (type === 'LineString') {
-                const line = this._computeLineStringGeometry(coordinates);
+            } else if (type === GeoJSONGeometryType.LINE_STRING) {
+                const line = this._geometryTransformer.computeLineString(coordinates);
                 geometries.push([line]);
-            } else if (type === 'MultiLineString') {
-                const multiline = this._computeMultiLineStringGeometry(coordinates);
+            } else if (type === GeoJSONGeometryType.MULTI_LINE_STRING) {
+                const multiline = this._geometryTransformer.computeMultiLineString(coordinates);
                 geometries.push(multiline);
-            } else if (type === 'Polygon') {
-                const polygon = this._computePolygonGeometry(coordinates);
+            } else if (type === GeoJSONGeometryType.POLYGON) {
+                const polygon = this._geometryTransformer.computePolygon(coordinates);
                 geometries.push([polygon]);
-            } else if (type === 'MultiPolygon') {
-                const multipolygon = this._computeMultiPolygonGeometry(coordinates);
+            } else if (type === GeoJSONGeometryType.MULTI_POLYGON) {
+                const multipolygon = this._geometryTransformer.computeMultiPolygon(coordinates);
                 geometries.push(multipolygon);
             }
         });
@@ -368,122 +300,19 @@ export default class GeoJSON extends Base {
         return geometries;
     }
 
-    _computePointGeometry (data) {
-        const lat = data[1];
-        const lng = data[0];
-        const wm = util.projectToWebMercator({ lat, lng });
-        return rsys.wToR(wm.x, wm.y, { scale: util.WM_R, center: this._center });
-    }
-
-    _computeLineStringGeometry (data, reverse) {
-        let line = [];
-        for (let i = 0; i < data.length; i++) {
-            const point = this._computePointGeometry(
-                data[reverse ? (data.length - i - 1) : i]
-            );
-            line.push(point.x, point.y);
-        }
-        return line;
-    }
-
-    _computeMultiLineStringGeometry (data) {
-        let multiline = [];
-        for (let i = 0; i < data.length; i++) {
-            let line = this._computeLineStringGeometry(data[i]);
-            if (line.length > 0) {
-                multiline.push(line);
-            }
-        }
-        return multiline;
-    }
-
-    _computePolygonGeometry (data) {
-        let polygon = {
-            flat: [],
-            holes: [],
-            clipped: []
-        };
-        let holeIndex = 0;
-        let firstReverse = false;
-
-        if (data.length) {
-            firstReverse = this._isReversed(data[0]);
-            const flat = this._computeLineStringGeometry(data[0], firstReverse);
-            polygon.flat = polygon.flat.concat(flat);
-        }
-        for (let i = 1; i < data.length; i++) {
-            if (firstReverse !== this._isReversed(data[i])) {
-                holeIndex += data[i - 1].length;
-                polygon.holes.push(holeIndex);
-            }
-            const flat = this._computeLineStringGeometry(data[i], firstReverse);
-            polygon.flat = polygon.flat.concat(flat);
-        }
-        return polygon;
-    }
-
-    _computeMultiPolygonGeometry (data) {
-        let multipolygon = [];
-        for (let i = 0; i < data.length; i++) {
-            let polygon = this._computePolygonGeometry(data[i]);
-            if (polygon.flat.length > 0) {
-                multipolygon.push(polygon);
-            }
-        }
-        return multipolygon;
-    }
-
-    _isReversed (vertices) {
-        let total = 0;
-        let pt1 = vertices[0];
-        let pt2;
-        for (let i = 0; i < vertices.length - 1; i++) {
-            pt2 = vertices[i + 1];
-            total += (pt2[1] - pt1[1]) * (pt2[0] + pt1[0]);
-            pt1 = pt2;
-        }
-        // When total is positive it means that vertices are oriented clock wise
-        // and, since positive orientation is counter-clock wise, it is reversed.
-        return total >= 0;
-    }
-
     _samplePoint (geometry) {
         const type = geometry.type;
 
         const coordinates = geometry.coordinates;
-        if (type === 'Point') {
+        if (type === GeoJSONGeometryType.POINT) {
             return coordinates;
-        } else if (type === 'LineString') {
+        } else if (type === GeoJSONGeometryType.LINE_STRING) {
             return coordinates[0];
-        } else if (type === 'MultiLineString' || type === 'Polygon') {
+        } else if (type === GeoJSONGeometryType.MULTI_LINE_STRING || type === GeoJSONGeometryType.POLYGON) {
             return coordinates[0][0];
-        } else if (type === 'MultiPolygon') {
+        } else if (type === GeoJSONGeometryType.MULTI_POLYGON) {
             return coordinates[0][0][0];
         }
-    }
-
-    // sets this._type, this._center, this._dataframeCenter
-    _setCoordinatesCenter () {
-        let x = 0;
-        let y = 0;
-        let nPoints = 0;
-        this._fetchFeatureGeometry({ sample: 10 }, (_, geometry) => {
-            if (!this._type) {
-                this._type = geometry.type;
-            }
-            const samplePoint = this._samplePoint(geometry);
-            const sampleXY = util.projectToWebMercator({ lng: samplePoint[0], lat: samplePoint[1] });
-            x += sampleXY.x;
-            y += sampleXY.y;
-            nPoints += 1;
-        });
-        if (nPoints > 1) {
-            x /= nPoints;
-            y /= nPoints;
-        }
-
-        this._center = { x, y };
-        this._dataframeCenter = rsys.wToR(this._center.x, this._center.y, { scale: util.WM_R, center: { x: 0, y: 0 } });
     }
 
     free () {