+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Next
+
+### Changed
+- Internal refactor regarding data sources (GeoJSON and Windshaft)
+
 ## [1.2.5] - 2019-04-25
 
 ### Changed